Separating specific noise definitions from general noise objects
