+++ pulser/simulation/noises.py
@@ -6,7 +6,6 @@ sampling time.
 """
 from __future__ import annotations
 
-import functools
 from typing import Callable, Optional
 
 import numpy as np
@@ -104,39 +103,3 @@ def doppler(reg: Register, std_dev: float, seed: Optional[int]) -> NoiseModel:
         )
 
     return f
-
-
-def compose_local_noises(*functions: NoiseModel) -> NoiseModel:
-    """Helper to compose multiple NoiseModel.
-
-    Args:
-        *functions: a list of functions
-
-    Returns:
-        The mathematical composition of *functions. The last element is applied
-        first. If *functions is [f, g, h], it returns f∘g∘h.
-    """
-    return functools.reduce(
-        lambda f, g: lambda x: f(g(x)), functions, lambda x: x
-    )
-
-
-def apply(
-    samples: list[QubitSamples], noises: list[NoiseModel]
-) -> list[QubitSamples]:
-    """Apply a list of NoiseModel on a list of QubitSamples.
-
-    The noises are composed using the compose_local_noises function, such that
-    the last element is applied first.
-
-    Args:
-        samples (list[QubitSamples]): A list of QubitSamples.
-        noises (list[NoiseModel]): A list of NoiseModel.
-
-    Return:
-        A list of QubitSamples on which each element of noises has been
-        applied.
-    """
-    tot_noise = compose_local_noises(*noises)
-
-    return [tot_noise(s) for s in samples]
